Mark classes final
